templates/langchain/translation_agent_builder/requirements.txt
langchain
langchain-openai
python-dotenv
openai